requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --resolver=backtracking requirements.in
#
aiobotocore==2.5.2
    # via s3fs
aiohttp==3.8.4
    # via
    #   aiobotocore
    #   s3fs
aioitertools==0.11.0
    # via aiobotocore
aiosignal==1.3.1
    # via aiohttp
asn1crypto==1.5.1
    # via
    #   oscrypto
    #   snowflake-connector-python
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
botocore==1.29.161
    # via aiobotocore
certifi==2023.5.7
    # via
    #   requests
    #   snowflake-connector-python
cffi==1.15.1
    # via
    #   cryptography
    #   snowflake-connector-python
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
    #   snowflake-connector-python
cryptography==40.0.2
    # via
    #   pyopenssl
    #   snowflake-connector-python
filelock==3.12.2
    # via snowflake-connector-python
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2023.6.0
    # via s3fs
greenlet==2.0.2
    # via sqlalchemy
idna==3.4
    # via
    #   requests
    #   snowflake-connector-python
    #   yarl
jinja2==3.1.2
    # via sqla-raw
jmespath==1.0.1
    # via botocore
markupsafe==2.1.3
    # via jinja2
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
oscrypto==1.3.0
    # via snowflake-connector-python
packaging==23.1
    # via snowflake-connector-python
pycparser==2.21
    # via cffi
pycryptodomex==3.18.0
    # via snowflake-connector-python
pyjwt==2.8.0
    # via snowflake-connector-python
pyopenssl==23.2.0
    # via snowflake-connector-python
python-dateutil==2.8.2
    # via botocore
pytz==2023.3
    # via snowflake-connector-python
requests==2.31.0
    # via snowflake-connector-python
s3fs==2023.6.0
    # via sqla-raw
six==1.16.0
    # via python-dateutil
snowflake-connector-python==3.0.4
    # via snowflake-sqlalchemy
snowflake-sqlalchemy==1.4.7
    # via -r requirements.in
sortedcontainers==2.4.0
    # via snowflake-connector-python
sqla-raw==1.5.0
    # via -r requirements.in
sqlalchemy==1.4.49
    # via
    #   snowflake-sqlalchemy
    #   sqla-raw
typing-extensions==4.7.1
    # via snowflake-connector-python
urllib3==1.26.16
    # via
    #   botocore
    #   requests
    #   snowflake-connector-python
wrapt==1.15.0
    # via aiobotocore
yarl==1.9.2
    # via aiohttp
